IMPALA-11964: Make sure Impala returns error for Iceberg tables with equality deletes
Impala only supports position deletes currently. It should raise an
error when equality deletes are encountered.

We already had a check for this when the query was planned by Iceberg.
But when we were using cached metadata the check was missing. This means
that Impala could return bogus results in the presence of equality
delete files. This patch adds check for the latter case as well.

Tables with equality delete files are still loadable by Impala, and
users can still query snapshots of it if they don't have equality
deletes.

Testing:
 * added e2e tests

Change-Id: I14d7116692c0e47d0799be650dc323811e2ee0fb
Reviewed-on: http://gerrit.cloudera.org:8080/19601
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>