IMPALA-11596: [DOCS] Document EXECUTE ROLLBACK functionality
Since IMPALA-11482 Impala supports the 'Alter Table Execute Rollback'
statement for Iceberg tables. Update the docs to cover this change.

The section on DESCRIBE HSTORY is expanded to include the output
columns, as this information is relevant to EXECUTE ROLLBACK.
The section on Cloning Iceberg tables is moved so that the sections
concerned with table history are adjacent.

TESTING:
- Built docs locally.

Change-Id: I0e1690378e560197263c49f468618b1ded922df3
Reviewed-on: http://gerrit.cloudera.org:8080/19606
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Noemi Pap-Takacs <npaptakacs@cloudera.com>
Reviewed-by: Zoltan Borok-Nagy <boroknagyz@cloudera.com>