IMPALA-11999: Ensure iceberg catalog location has correct filesystem prefix for Ozone
Some iceberg tests run using different forms of iceberg catalog. One of
these specifies iceberg.catalog_location as one of the TBLPROPERTIES
when a table is created. This location was hard-coded to a path
starting with "/test-warehouse", which results in an illegal Ozone
path. Fix this by wrapping the location with get_fs_path(), which gives
a legal Ozone path.

TESTING
- ran end to end tests on an Ozone filesystem.

Change-Id: Iabafb9cb778a923ffe084dd3e61053b2aa8a6b1f
Reviewed-on: http://gerrit.cloudera.org:8080/19629
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>