IMPALA-11998: Fix potential invalid iterator issue in ImpalaServer::GetQueryRecord()
The current ImpalaServer::GetQueryRecord() may cause the caller to
access invalid iterators, although the function locks query_log_lock_
in the execution, the query_record it provides cannot guarantee to be
valid, because it is out of the protection of query_log_lock_ after
returning, if the corresponding record in query_log_index_ is just
deleted at this time, then query_record will be an invalid iterator.

To fix that, this patch uses shared_ptr instead of the original
unique_ptr to extend the lifetime of query_record after GetQueryRecord
returns. In addition, a new argument retried_query_record is added to
GetQueryRecord(), the motivation is to acquire query_log_lock_ only once
for getting both the original query record and the retried query record.

Change-Id: I1279558b0d0b2a3ba31a496c0acb84ac9d99f2f1
Reviewed-on: http://gerrit.cloudera.org:8080/19621
Reviewed-by: Quanlong Huang <huangquanlong@gmail.com>
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>