IMPALA-11822: (Addendum) Optimize the Refresh/Invalidate event processing by skipping unnecessary events
IMPALA-11822 introduced a new argument 'lastRefreshEventId' to the
HdfsPartition constructor and this is causing Hive builds to fail.
This addendum patch addresses this issue by keeping the old constructor
in the HdfsPartition class and call the new constructor with -1L as
default value for 'lastRefreshEventId'.

Testing: Verified that the Hive builds are passing with this patch.

Change-Id: I803c13b58e3de047c12f75a584232819c7dc74c1
Reviewed-on: http://gerrit.cloudera.org:8080/19624
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>