IMPALA-12047: [DOCS] Fix a URL mistake when using "git clone" in docs/README.md
There's a URL mistake when using "git clone" in docs/README.md:
  git clone https://gitbox.apache.org/repos/asf/impala.git/docs

This doesn't work and it will prompt "repository not found".

This change corrects the description, providing two ways to download
the docs - either by downloading the whole repository and going to the
docs/ directory or by downloading only the docs using git sparse-checkout.

Change-Id: Ib00c37e28e67cca5b3630742b4c366dea4e967b7
Reviewed-on: http://gerrit.cloudera.org:8080/19634
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Daniel Becker <daniel.becker@cloudera.com>
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>