IMPALA-12014 warning on failed KeepAlive for Kudu scanner
With this patch, there is now a warning message output when
ScannerKeepAlive RPC for a Kudu scanner fails.  To avoid flooding the
log with those messages, KLOG_EVERY_N_SECS(WARNING, 60) is used.

I also updated the signature of the KuduScanner::BuildErrorString()
method to become a constant one.

This is a follow-up to IMPALA-3292.

Change-Id: If39f968685797506491d3d5ebc481efb6e43a568
Reviewed-on: http://gerrit.cloudera.org:8080/19643
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Michael Smith <michael.smith@cloudera.com>