IMPALA-12028: test_execute_rollback might fail due to used timezone
There are some steps in the test that set the timezone and execute
rollback to given time. However, the snapshot creation, and querying
the current time uses local timezone. As a reuslt if the test is run
e.g. in CET timezone it fails when expecting an error in the Icelandic
timezone.

Tests:
  - Re-ran TestIcebergTable.test_execute_rollback.

Change-Id: Iba9724f9b86cc508e6497eb33844a6480498b6e4
Reviewed-on: http://gerrit.cloudera.org:8080/19655
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>