IMPALA-12038: Switch report_benchmark_results.py to python 3
When using TPC-DS with a large number of iterations, the
results JSON files are enormous. Using Python2,
report_benchmark_results.py runs out of memory and fails to
produce the report. Python 3 is more efficient in how it
processes Unicode inputs (see Python PEP-0393), so it's
memory usage is much lower. It is able to handle generating
reports that Python 2 cannot.

As a general cleanup, this fixes all the flake8 issues for this file.

Testing:
 - Processed very large JSON results (4+GB each for both baseline
   result and new result). Python 3 completes successfully when
   Python 2 failed.

Change-Id: Idbde17f720b18d38dc2c2104ecf3fec807c1839d
Reviewed-on: http://gerrit.cloudera.org:8080/20918
Reviewed-by: Riza Suminto <riza.suminto@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>