IMPALA-11970: Optimized rendering for the query timeline display
Decoupled the retrieval of attributes and rendering
to optimize window resize events for large JSON
query profiles in the timeline display.

A reload will be triggered after every second,
in case the query execution has not completed,
or due to errors in parsing the profile;
only valid profiles are rendered.

Replaces 'maximum()' and 'minimum()' functions with
'Math.max.apply()' and 'Math.min.apply()'.

Change-Id: Iec458602266863b6e313f52b26a2dc35cff0db0f
Reviewed-on: http://gerrit.cloudera.org:8080/19744
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>