IMPALA-12055: Use jquery.min.js from local repo
Our website depends on
https://ajax.googleapis.com/ajax/libs/jquery/1.11.3/jquery.min.js
In some regions, it's pretty slow to get this file and it usually fails
at last.

This adds the file to the asf-site branch so the website can use a local
copy of it. Also bumps the jquery version to 3.5.1 to match the master
branch.

Tests:
 - Verified the pages locally.

Change-Id: Ie8c4a9c2b9b64128e610a803363c2c5be6e57018
Reviewed-on: http://gerrit.cloudera.org:8080/19720
Reviewed-by: Daniel Becker <daniel.becker@cloudera.com>
Tested-by: Quanlong Huang <huangquanlong@gmail.com>