IMPALA-12058: Impala throw exception for CTAS with non unique primary key
The auto-incrementing column is added to target table by Impala frontend
for CTAS statement when creating Kudu table with non unique primary key
so that the Impala internal table has same layout as the table created
by Kudu engine. This is required for insertion operation of CTAS.
But auto-incrementing column is not in expression. It causes Planner
throws exception when computing Lineage graph for the column.

This patch fixes the issue by omiting auto-incrementing column for Kudu
table when adding target column labels for computing Lineage graph.

Testing:
 - Added end-to-end test case for creating Kudu tables with Lineage
   enabled.
 - Passed core tests.

Change-Id: Ic87b7fac8eb857e9dc5c59bdb448270b23f6dd5a
Reviewed-on: http://gerrit.cloudera.org:8080/19725
Reviewed-by: Kurt Deschler <kdeschle@cloudera.com>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Reviewed-by: Riza Suminto <riza.suminto@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>