IMPALA-12063: Upgrade zlib to 1.2.13
This bumps the toolchain version to get zlib 1.2.13,
which contains the fix for CVE-2022-37434.

This toolchain incorporates several changes to clean
up the native-toolchain and remove unnecessary
component builds. As part of this, OpenSSL is no
longer built in the toolchain, so this stops
downloading it. This changes the build to require
OpenSSL 1.0.2 or higher. This doesn't impact anything,
because all supported platforms already used
OpenSSL 1.0.2 or higher. See IMPALA-12064.

Testing:
 - perf-AB-test shows no change in performance for
   ORC with deflate
 - GVO passes

Change-Id: I96efc947534cda8d15d4f440cd6851d397b6562d
Reviewed-on: http://gerrit.cloudera.org:8080/19760
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>