IMPALA-12076: Use ZSTD interfaces with reusable context
For repeated compression/decompression, ZSTD recommends
constructing a context once via ZSTD_createCCtx()/ZSTD_createDCtx()
and using the set of interfaces that passes in the context explicitly
to avoid constructing the context on each call.

This follows the recommendation and allocates the ZSTD context once for
each compressor / decompressor and reuses it for the lifetime of the
compressor / decompressor.

This gets a minor speedup for small-scale ZSTD TPC-H:
+----------+------------------------+---------+------------+------------+----------------+
| Workload | File Format            | Avg (s) | Delta(Avg) | GeoMean(s) | Delta(GeoMean) |
+----------+------------------------+---------+------------+------------+----------------+
| TPCH(42) | parquet / zstd / block | 3.55    | -1.40%     | 2.52       | -1.63%         |
+----------+------------------------+---------+------------+------------+----------------+

Testing:
 - Ran core job
 - Ran a perf-AB-test job

Change-Id: I5010a56bf8202ccb3f1710425002f81587fd412b
Reviewed-on: http://gerrit.cloudera.org:8080/19773
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>