IMPALA-12082: Fix db not found error of INVALIDATE METADATA under unloaded db
INVALIDATE METADATA can be executed on tables under unloaded db. It will
bring up the metadata of the db by the way. However, this feature is
broken after IMPALA-11808 in which we try to get the table from catalog
cache assuming it's loaded. This causes the above use case failed by
DatabaseNotFoundException.

This patch fixes the regression by not getting the table from catalog
cache for INVALIDATE METADATA commands. We only do so for REFRESH
commands. After the INVALIDATE METADATA command succeeds, if we need to
fire reload events, we get the table from catalog cache.

Tests:
 - Add e2e tests for event-processor is disabled and enabled.

Change-Id: Ifd0a9e87f06c38f569c32bd10cc2668403681fd4
Reviewed-on: http://gerrit.cloudera.org:8080/19786
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>