IMPALA-12084: IM under unloaded db should not send full db updates in catalog-v2
INVALIDATE METADATA under unloaded db will also add the newly added db
along with the newly added table in the catalog update. In LocalCatalog
mode, we should only send back invalidation of the db instead of the
full db object. This is a missing logic we should have added in
IMPALA-9936.

Tests:
 - Add e2e test on LocalCatalog mode

Change-Id: I38a44f69cca20331ab31acc270c2921a7bfedae5
Reviewed-on: http://gerrit.cloudera.org:8080/19787
Reviewed-by: Daniel Becker <daniel.becker@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Michael Smith <michael.smith@cloudera.com>