IMPALA-12086: Fix BitUtil::CountLeadingZeros for zero
BitUtil::CountLeadingZeros() uses _builtin_clz() and _builtin_clzll() to
return the number of leading zeros in a number. Calling these functions
with 0 is undefined.

This change introduces a check in CountLeadingZeros() to only call these
functions if the argument is non-zero.

Also makes CountLeadingZeros() constexpr and restricts it to type sizes
that can actually be handled by the function.

Testing:
 - Added a test for CountLeadingZeros() in bit-util-test.cc

Change-Id: I0ae3d84c5c25871388155747f91c51162c1a0590
Reviewed-on: http://gerrit.cloudera.org:8080/19792
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>