IMPALA-12096: Add process start time and version to backends webui
This patch adds process start time and version to the /backends page.

Two more optional elements are added in BackendDescriptorPB and can
be broadcast through statestore topic. This information should be
helpful for users checking all backends in a large cluster.

For display, as two more columns are added to the table of backend
information, the table is changed to 'table-responsive' to be scrolled
horizontally with ease. A sample screenshot is attached to the
IMPALA-12096 ticket.

Testing:
- Added cases to test_web_pages.py

Change-Id: I5f1f0ba0081986f428840442c247d7dde9e1ba05
Reviewed-on: http://gerrit.cloudera.org:8080/19800
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>